Fix: connection key + placement remove account

+++ html/home/js/connection.js
@@ -4,6 +4,17 @@ jQuery(document).ready(function ($) {
     if (sessionStorage.getItem('apikey')) {
         connected_func()
     }
+
+    document.getElementById("pseudo").addEventListener("keydown", function(event) {
+        if (event.key === "Enter" || event.key === 13) {
+            document.getElementById("btn_connection").click();
+        }
+    });
+    document.getElementById("password").addEventListener("keydown", function(event) {
+        if (event.key === "Enter" || event.key === 13) {
+            document.getElementById("btn_connection").click();
+        }
+    });
 });
 
 const connection = async function () {
@@ -11,38 +22,46 @@ const connection = async function () {
     if (sessionStorage.getItem('apikey')) {
         connected_func()
     } else {
-        await $.ajax({
-            type: 'GET',
-            url: '/php/apikey.php',
-            dataType: 'json',
-            success: function(data) {
-                sessionStorage.setItem("masterkey", data.masterkey);
-            }
-        });
-        await $.ajax({
-            type: 'PUT',
-            url: "/api/v1/identity/connection/",
-            headers: {
-                "APIKEY": sessionStorage.getItem("masterkey"),
-                'Accept': 'application/json',
-                "Content-Type": "application/json"
-            },
-            data: JSON.stringify({
-                "pseudo": $('#pseudo').val().toLowerCase(),
-                "password": $('#password').val()
-            }),
-            success: function (data) {
-                sessionStorage.clear();
-                sessionStorage.setItem("apikey", data.APIKEY);
-                sessionStorage.setItem('pseudo', data.pseudo);
-                connected_func()
-            },
-            error: function(data) {
+        try {
+            await $.ajax({
+                type: 'GET',
+                url: '/php/apikey.php',
+                dataType: 'json',
+                success: function(data) {
+                    sessionStorage.setItem("masterkey", data.masterkey);
+                }
+            });
+            await $.ajax({
+                type: 'PUT',
+                url: "/api/v1/identity/connection/",
+                headers: {
+                    "APIKEY": sessionStorage.getItem("masterkey"),
+                    'Accept': 'application/json',
+                    "Content-Type": "application/json"
+                },
+                data: JSON.stringify({
+                    "pseudo": $('#pseudo').val().toLowerCase(),
+                    "password": $('#password').val()
+                }),
+                success: function (data) {
+                    sessionStorage.clear();
+                    sessionStorage.setItem("apikey", data.APIKEY);
+                    sessionStorage.setItem('pseudo', data.pseudo);
+                    connected_func()
+                },
+                error: function(data) {
+                    sessionStorage.clear();
+                    $('#loader_connection').css("visibility", "hidden");
+                    $('.connection_input').addClass("vibration_input");
+                }
+            });
+        } catch (error) {
+            if (400 > error.status || error.status > 499) {
                 sessionStorage.clear();
                 $('#loader_connection').css("visibility", "hidden");
                 $('.connection_input').addClass("vibration_input");
             }
-        });
+        }
     }
 }
 
@@ -56,17 +75,19 @@ const connected_func = function(){
 }
 
 const disconnected_func = function() {
-    $('#loader_connection').css("visibility", "visible");
-    setTimeout(() => {
-        localStorage.clear();
-        sessionStorage.clear();
-        $('.connection_input').removeAttr("style").val('');
-        $('.connection_button').removeAttr("style");
-        $('.connection_off').removeAttr("style");
-        $('.connection_on').hide();
-        $('.connection_input').removeClass("vibration_input")
-        $('#loader_connection').css("visibility", "hidden");
-    }, 1500);
+    return new Promise((resolve, reject) => {
+        $('#loader_connection').css("visibility", "visible");
+        setTimeout(() => {
+            localStorage.clear();
+            sessionStorage.clear();
+            $('.connection_input').removeAttr("style").val('');
+            $('.connection_button').removeAttr("style");
+            $('.connection_off').removeAttr("style");
+            $('.connection_on').hide();
+            $('.connection_input').removeClass("vibration_input")
+            $('#loader_connection').css("visibility", "hidden");
+        }, 1000);
+    });
 }
 
 const format_date = function(date){
